Added check to verify running as root.

+++ deploy-environment.sh
@@ -2,28 +2,37 @@
 
 # Usage: Run on vagrant machine or another Debian box as root (not sudo - actually switch to root):
 # bash <(curl -s https://raw.githubusercontent.com/ITMT-430/Team-3-Install-Scripts/master/deploy-environment.sh)
-if test -f "/bs/eucainstall";
-then
-  apt-get update
-  apt-get install -y git
-  apt-get install -y python-setuptools python-dev libxslt1-dev libxml2 libxml2-dev zlib1g-dev
-  apt-get remove python-requests
-  git clone https://github.com/jhajek/euca2ools.git /euca2ools
-  cd /euca2ools
-  git checkout origin/maint-3.1
-  python setup.py install
-  read -p "Verify there are no errors up to this point, then press any key to continue." nothing
+if [ $USER = "root" ]; then
+  if test -f "/bs/eucainstall";
+  then
+    apt-get update
+    apt-get install -y git
+    apt-get install -y python-setuptools python-dev libxslt1-dev libxml2 libxml2-dev zlib1g-dev
+    apt-get remove python-requests
+    git clone https://github.com/jhajek/euca2ools.git /euca2ools
+    cd /euca2ools
+    git checkout origin/maint-3.1
+    python setup.py install
+    read -p "Verify there are no errors up to this point, then press any key to continue." nothing
+    clear
+    read -p "Copy your credential file to /euca2ools/creds.zip, then press any key to continue." nothing
+    unzip -d creds creds.zip
+    source /euca2ools/creds/eucarc
+    echo "source /euca2ools/creds/eucarc">>~/.bashrc
+    echo "source /euca2ools/creds/eucarc">>~/.zshrc
+    mkdir /bs
+    touch /bs/eucainstall
+  fi
+  mainmenu
+else
   clear
-  read -p "Copy your credential file to /euca2ools/creds.zip, then press any key to continue." nothing
-  unzip -d creds creds.zip
-  source /euca2ools/creds/eucarc
-  echo "source /euca2ools/creds/eucarc">>~/.bashrc
-  echo "source /euca2ools/creds/eucarc">>~/.zshrc
-  mkdir /bs
-  touch /bs/eucainstall
+  echo Hey you!  Ya... you!  You''re an idiot!
+  echo Read the freaking documentation!
+  echo You''re supposed to run this as root!!!
+  read -p "Go learn how to computer." nothing
 fi
 
-mainmenu
+
 
 function mainmenu {
   clear
